Separate processor from web api in the codebase
