+++ .dockerignore
@@ -1,0 +1,2 @@
+scripts/minio-conf/
+scripts/minio-data/
