+++ manage.py
@@ -2,7 +2,7 @@ from flask_migrate import Migrate
 from flask_migrate import MigrateCommand
 from flask_script import Manager
 
-from ros.app import app, db
+from ros.lib.app import app, db
 
 migrate = Migrate(app, db)
 manager = Manager(app)
