--- run.py
@@ -1,4 +1,0 @@
-from ros.app import app
-
-if __name__ == "__main__":
-    app.run(host='0.0.0.0', port=8080)
