+++ scripts/.env
